fix circular dependency in club-supervisor relation
