+++ backend/db/schema/index.ts
@@ -2,3 +2,5 @@ export * from './user';
 export * from './club';
 export * from './event';
 export * from './ticket';
+export * from './supervisor';
+export * from './relations';
